requirements.txt
python-binance
colorama
aiohttp
